+++ packages/band/src/dendrogram/index.tsx
@@ -1,0 +1,6 @@
+export * from './context'
+export * from './Dendrogram'
+export * from './DendrogramLabels'
+export * from './DendrogramTree'
+export * from './predicates'
+export * from './types'