+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Added
 
+-   [band] components `Dendrogram`, `DendrogramTree` to summarize hierarchical
+    dendrograms
 -   [core] linear scale specs with and without the `nice` parameter
 -   [all] more keywords in `package.json` files